refactor:#265 루트 리팩토링
