+++ src/app/(auth)/change-password/_components/step-current-password.tsx
@@ -8,12 +8,12 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
-import formSchema from '@/app/schemas/form-schema.schema';
+import formSchema from '@/schemas/form-schema.schema';
 import { Typography } from '@/components/ui/typography';
 import SITE_MAP from '@/constants/site-map.constant';
 import PUBLIC_ERROR_MESSAGE from '@/constants/public-error-message.constant';
 import { signIn } from '@/apis/auth-server.api';
-import { useUserStore } from '@/store/user-store';
+import { useUserStore } from '@/store/user.store';
 
 const passwordSchema = z.object({
   password: formSchema.PASSWORD_SCHEMA
