+++ src/app/(auth)/change-password/_components/step-new-password.tsx
@@ -9,13 +9,13 @@ import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
 import { changePassword, signIn } from '@/apis/auth-server.api';
-import formSchema from '@/app/schemas/form-schema.schema';
+import formSchema from '@/schemas/form-schema.schema';
 import { Typography } from '@/components/ui/typography';
 import SITE_MAP from '@/constants/site-map.constant';
 import PUBLIC_ERROR_MESSAGE from '@/constants/public-error-message.constant';
 import Modal from '@/components/commons/modal';
 import { StepNewPasswordType } from '../_types/funnel.type';
-import { useUserStore } from '@/store/user-store';
+import { useUserStore } from '@/store/user.store';
 
 const passwordSchema = z.object({
   password: formSchema.PASSWORD_SCHEMA
