+++ src/app/(auth)/sign-up/_components/step-password.tsx
@@ -4,7 +4,7 @@ import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
-import formSchema from '@/app/schemas/form-schema.schema';
+import formSchema from '@/schemas/form-schema.schema';
 import { useForm } from 'react-hook-form';
 import { useEffect, useRef } from 'react';
