+++ src/app/(client)/meal/post/edit/_components/add-meal-card-drawer.tsx
@@ -9,7 +9,7 @@ import { Button } from '@/components/ui/button';
 import IconButton from '@/components/commons/icon-button';
 import { Typography } from '@/components/ui/typography';
 import MealPostAddMealAiLoading from '@/app/(client)/meal/post/_components/add-meal-ai-loading';
-import { useUserStore } from '@/store/user-store';
+import { useUserStore } from '@/store/user.store';
 import { generateCaloriesAnalysisByText } from '@/apis/gemini.api';
 import { parseGeminiResponse } from '@/lib/gemini';
 import { AiResponseDTO } from '@/types/DTO/ai_analysis.dto';
