+++ src/app/(client)/set-goal/_components/step-purpose.tsx
@@ -4,7 +4,7 @@ import { Typography } from '@/components/ui/typography';
 import { PurposeType } from '@/types/user-personal-info.type';
 import { ChangeEvent, FormEvent, useState } from 'react';
 import { GOAL_OPTIONS } from '../_constants/funnel.constant';
-import { useUserStore } from '@/store/user-store';
+import { useUserStore } from '@/store/user.store';
 
 type StepPurposeProps = {
   nextStep: (data: PurposeType) => void;
