+++ src/app/(home)/_components/home-content.tsx
@@ -4,7 +4,7 @@ import GlassBackground from '@/components/commons/glass-background';
 import { getMyMealByDate } from '@/apis/meal.api';
 import { useDateContext } from '@/app/(home)/_contexts/date.context';
 import { calculateTotalNutrition } from '@/utils/nutrition-calculator.util';
-import { useUserStore } from '@/store/user-store';
+import { useUserStore } from '@/store/user.store';
 import { MealDTO } from '@/types/DTO/meal.dto';
 import { useEffect, useState } from 'react';
 import AiFeedbackText from './ai-feedback-text';
